requirements.txt
streamlitpandasbeautifulsoup4lxml